update voor wallet
